+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fixed an issue where DIM Sync data might not be available when Bungie.net is down.
 * Added `source:rootofnightmares`/`source:ron` and `modslot:rootofnightmares` searches.
 * DIM now correctly allows you to unsave previously saved queries that later became invalid.
+* Fixed the `is:curated` filter never matching weapons without an equipped kill tracker.
 
 ## 7.60.1 <span class="changelog-date">(2023-03-14)</span>